requirements.txt
dbt-bigquery
requests
pandas
apache-airflow
pyarrow
google-cloud-storage